fix the extraction of translatable strings
closes #175
