+++ liberapay/renderers/jinja2.py
@@ -1,8 +1,43 @@
-import aspen_jinja2_renderer as base
+from functools import wraps
 
+import aspen_jinja2_renderer as base
+from jinja2 import Undefined
 from markupsafe import escape as htmlescape
 
-from liberapay.website import JINJA_ENV_COMMON
+from ..i18n.extract import JINJA_BASE_OPTIONS
+from ..website import website
+
+
+def wrap_method(method):
+    @wraps(method)
+    def f(self, *a, **kw):
+        try:
+            self._fail_with_undefined_error()
+        except Exception as e:
+            website.tell_sentry(e, level='warning')
+        return method(self, *a, **kw)
+    return f
+
+
+class CustomUndefined(Undefined):
+    """This subclass sends errors to Sentry instead of actually raising them.
+
+    Doc: https://jinja.palletsprojects.com/en/2.11.x/api/#undefined-types
+    """
+    __iter__ = wrap_method(Undefined.__iter__)
+    __str__ = wrap_method(Undefined.__str__)
+    __len__ = wrap_method(Undefined.__len__)
+    __eq__ = wrap_method(Undefined.__eq__)
+    __ne__ = wrap_method(Undefined.__ne__)
+    __bool__ = wrap_method(Undefined.__bool__)
+    __hash__ = wrap_method(Undefined.__hash__)
+
+
+JINJA_ENV_COMMON = dict(
+    JINJA_BASE_OPTIONS,
+    auto_reload=website.env.aspen_changes_reload,
+    undefined=CustomUndefined,
+)
 
 
 class Renderer(base.Renderer):
